pyproject.toml
[project]
name = "Learn_AI_Agents_workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.24.0",
    "openai>=1.93.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
